requirements.txt
pluginbase==1.0.0
loguru==0.5.3
pyyaml==5.4.1
pyhelm==2.14.5
deepdiff==5.8.1
yamllint==1.26.0
pylint==2.15.5